Failure detector refactoring. Fix tests
