+++ test/src/test/resources/logback-test.xml
@@ -34,7 +34,7 @@
     <!-- Control logging levels for individual components here. -->
     <logger name="org.corfudb.runtime.object" level="INFO"/>
     <logger name="org.corfudb.runtime.clients" level="INFO"/>
-    <logger name="org.corfudb.infrastructure" level="debug"/>
+    <logger name="org.corfudb.infrastructure" level="info"/>
     <logger name="io.netty.util" level="INFO"/>
     <logger name="io.netty.util.internal" level="INFO"/>
     <logger name="io.netty.buffer" level="INFO"/>
@@ -44,9 +44,9 @@
     </logger>
 
 
-    <root level="debug">
+    <root level="info">
         <!--<appender-ref ref="FILE" />-->
-        <appender-ref ref="STDOUT" />
+        <!--<appender-ref ref="STDOUT" />-->
         <!--<appender-ref ref="MetricsRollingFile" />-->
     </root>
 </configuration>
